+++ doc/CHANGES.rst
@@ -17,6 +17,7 @@ CHANGELOG
 
 * [FIX] check constraint must not be create at the same time that the column, 
   because the column of the constraint could have not exist yet
+* [REF] change namimg convention
 
 0.12.2 (2018-01-04)
 -------------------